+++ prisma/migrations/20250205210027_added_fullname/migration.sql
@@ -1,0 +1,10 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `name` on the `Account` table. All the data in the column will be lost.
+  - Added the required column `fullname` to the `Account` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- AlterTable
+ALTER TABLE "Account" DROP COLUMN "name",
+ADD COLUMN     "fullname" TEXT NOT NULL;