+++ prisma/migrations/20250206000905_draft_added_name/migration.sql
@@ -1,0 +1,9 @@
+/*
+  Warnings:
+
+  - Made the column `name` on table `Account` required. This step will fail if there are existing NULL values in that column.
+
+*/
+-- AlterTable
+ALTER TABLE "Account" DROP COLUMN "name",
+ADD COLUMN  "biography" TEXT NOT NULL;